[project @ 2000-11-01 11:41:47 by simonmar] Add a basic "front panel" for GHC-compiled programs.
How to use it:

	- re-autoconf & configure to detect GTK+

	- add "GhcRtsWithFrontPanel = YES" to mk/build.mk

	- rebuild the RTS

	- compile up a program, add `gtk-config --libs` to the
	  link command line

	- run with program with +RTS -f,

	- sit back & watch the show :-)  Programs with lots of
	  heap-resident data are the most interesting.  For extra
	  kicks, turn up the number of generations & steps like so:
	  +RTS -f -G5 -T3.

	- Bootstrap your compiler, and see in glorious technicolor
	  just how much of a lumbering beast GHC really is.

This is a work in progress.  There's lots more stuff we could display
on the panel: suggestions/comments are of course welcome.  The window
layout was designed with GLADE, I'll commit the config file shortly.

I haven't quite figured out how we're going to integrate this with the
release yet (ie. whether we'll distribute two separate RTS's or what).