[project @ 2000-11-06 13:13:17 by simonmar] Change the way source distributions are built.
Now we do a 'make boot', 'make distclean', build a link tree to filter
some files out (eg. .cvsignore), and finally tar up the result.  The
process isn't bulletproof, so it makes sense to compare the manifest
against previous versions.

Now you don't have to be in a linked build tree to make a source
distribution.