[project @ 2001-01-03 11:18:51 by simonmar] s/boxed/lifted/
The typechecker's notion of "boxed" versus "unboxed" kind should
really have been "unlifted" versus "lifted" instead.  It is illegal to
unify an unlifted (but boxed) type with a polymorphic type variable,
since an unlifted/boxed type is always assumed to be a pointer to the
object itself, never a thunk or indirection.

This commit removes isUnboxedType, and renames a bunch of things that
were previously boxed/unboxed to unlifted/lifted.