[project @ 2001-01-18 16:30:00 by simonmar] Problem 1 =========
The typechecker, when deciding whether to extend the Package
environment with any new typechecked declarations in its hand, was
inserting new declarations into the environment only if the
declaration's module differed from the "current" module.  This doesn't
work if the "current" module is a package module, as it could be at
the GHCi command line, for example.

The solution is to filter the declarations only if the current module
is not a package module.


Problem 2
=========

The "current" module, as obtained from the compilation manager, was
always bogusly a Home module (it used mkHomeModule).  To properly fix
this, the GHCi state has to carry around Modules instead of
ModuleNames, and CompMan.cmLoadModule needs to return a list of
Modules.