[project @ 2001-01-26 17:21:51 by simonmar] Greatly simplify the story about linkables, source_unchanged, and the pre-upsweep.
Now we pre-generate the list of valid linkables; that is, for each
module if a linkable exists and is newer than the source, we keep it.
If a module has a valid linkable, then it is "source unchanged", and
it is also possibly "stable" as far as the pre-upsweep is concerned
(as long as its imports are also stable).

The pre-upsweep is no longer dependent on the mode
(interactive/batch).

There's still a bug here, though: the pre-upsweep removes old
interfaces from the HIT, so we don't get an opportunity to avoid
compilation for non-stable modules.  That's the next job.