[project @ 2001-02-22 13:17:57 by simonpj] fromInt
Remove fromInt from class Num, though it is retained
as an overloaded operation (with unchanged type) in PrelNum.

There are quite a few consequential changes in the Prelude.
I hope I got them all correct!

Also fix a bug that meant Integer (and its instances)
wasn't getting slurped in by the renamer, even though it
was needed for defaulting.