[project @ 2001-02-27 11:50:05 by simonpj] Temporary fix for a nasty black hole
The problem is that the type checker has a big knot for "unf_env".
This means that we can't look at unfoldings inside the loop, which
is fair enough.  But setting an unfolding in the IdInfo is strict
in the unfolding, so we can't look at the IdInfo either.

But isLocalId looks at the IdInfo, and it was being used in an
assert in TcHsSyn, and in setting the in_scope_vars in TcIfaceSig.

I think the right solution is to take the "flavour" out of IdInfo,
and put it into VarDetails, but I've done a quick fix for now.
(Remove the assert, and use a different way in TcIfaceSig.)