[project @ 2001-02-28 17:57:52 by simonmar] Compiling a module outside of a running GHCi and expecting GHCi to pick it up when you do ':r' ain't gonna work.
We thought it would, but there's a problem with module versions and
the recompilation machinery, namely that the outside compiler doesn't
know about the versions being used internally in the running GHCi.

The "right" solution is to implement a ':compile' command in GHCi, but
we aren't going to do that before the release.

This commit disables the picking up of newly compiled modules, until
you do a full load (i.e. :load).