[project @ 2001-03-02 14:28:44 by simonmar] Yet another good bug (I'm on a roll today! :-)
When squeezing update frames, make sure the updatee pointers aren't
identical before updating one to point to the other.  This is mostly
harmless, but in one example I have caused the program to go into an
infinite loop rather than fall down a black hole.

To be merged into the 4.08 branch.