[project @ 2001-03-19 16:13:22 by simonpj] ------------------------------- 	Improve orphan-module resolution 	-------------------------------
Consider the following rule (and there are lots of these in
the Prelude):

	fromIntegral T = fromIntegral_T

where T is defined in the module being compiled.

is an orphan.  Of course it isn't, an declaring it an orphan would
make the whole module an orphan module, which is bad.

This commit arranges to determine orphan rules, and the orphan-hood
of a module, much later than before.  (Before mi_orphan was set by
the renamer, now it is set by MkIface.)