[project @ 2001-03-28 05:07:34 by kglynn] Cpr needs the correct arity for an imported function.  In some cases the arity reported in the interface can be too small.
In this test case the function is returning a newtype,  and the newtype
is hiding a function.  The arity in the interface file says 0.  It is
triggered by compiling the cpr001.hs test files with profiling on. The code
comes from a bug report.

I should look for a better case,  because the profiling flaw which reveals
this problem will hopefully go away some time.