[project @ 2001-04-05 11:31:26 by simonpj] -------------------------------- 	Better grouping for ty/cls decls 	--------------------------------
When finding mutually-recursive groups of type and class decls,
we shouldn't include classes mentioned in a deriving clause as
edges. E.g.

	class Eq a where ...

	data Bool = True | False deriving( Eq )

Eq and Bool are not mutually recursive.

The edges are computed by RnHsSyn.tyClDeclFVs, so we remove the
derivings from there.

There a consequential fix in RnSource.rnSourceDecl