[project @ 2001-04-06 16:15:39 by sewardj] Commit the following change:
> > Why the default libdir is /usr/local/lib
> > and not /usr/local/lib/ghc-<version>?
>
> Great question.  I end up running config with the likes of
> `--libdir=/usr/local/lib/ghc-5.0' every time, which gets to
> be annoying ;-)

I've been meaning to fix this for a while, but couldn't see a good way
to do it.  I found a (mildly-hacky) way to do it today: in
fptools/ghc, we override $(libdir) to be
$(libdir)/$(ProjectNameShort)-$(ProjectVersion), so everything inside
fptools/ghc will be installed in the subdirectory.  fptools/hslibs is
a bit more of a hack.