[project @ 2001-05-03 12:33:50 by simonpj] ****	MERGE WITH 5.00 BRANCH     ********
	--------------------------------
	Monomorphism restriction for implicit parameters
	--------------------------------

This commit tidies up the way in which monomorphic bindings
are dealt with, incidentally fixing a bug to do with implicit
parameters.

The tradeoffs concerning monomorphism and implicit paramters are
now documented in TcSimplify.lhs, and all the strategic choices
are made there (rather than in TcBinds where they were before).

I've continued with choice (B) -- which Jeff first implemented --
because that's what Hugs does, lacking any other consensus.