[project @ 2001-05-04 14:43:26 by simonpj] ****	MERGE WITH 5.00 BRANCH     ********
	--------------------------------
	Fix a black hole when type checking type decls
	--------------------------------

GHC was falling into a black hole when type checking a recursive
group of type declarations including a chain of type synonyms.

  type PhraseFun = PMap -> Float
  type PMap      = () -> Player
  data Player    = P.MkT P.PhraseFun

Reason: too much consistency checking in TcMonoType.
Easily fixed using the existing wimp_out hack, but it's a mess.
This commit fixes it for the 5.00 branch but I'll do something
better in the head shortly.